requirements.txt
streamlit
pillow
easyocr
openai>=1.0.0
azure-storage-blob
numpy
torch
opencv-python-headless
requests
azure-cognitiveservices-vision-computervision
msrest


